+++ resources/views/roles/index.blade.php
@@ -7,10 +7,12 @@
             <div class="card">
                 <div class="card-header">
                     Role List
-                    <a href="{{route('role.create')}}">
+                    @can('create_role')
+                        <a href="{{route('role.create')}}">
 
-                        <button type="button" class="btn btn-link role-create"><i class="fa fa-plus-circle" aria-hidden="true"></i>&nbsp;&nbsp;Create New Role</button>
-                    </a>
+                            <button type="button" class="btn btn-link role-create"><i class="fa fa-plus-circle" aria-hidden="true"></i>&nbsp;&nbsp;Create New Role</button>
+                        </a>
+                    @endcan
                 </div>
 
                 <div class="card-body">
@@ -35,17 +37,20 @@
                                         <tr>
                                             <td>{{$item->name}}</td>
                                             <td class="row">
-                                                {!! Form::open(['route'=>'role.edit','method'=>'get']) !!}
-                                                {{ Form::hidden('id',$item->id) }}
-                                                <button class="btn btn-xs btn-success">Edit</button>
+                                                @can('create_role')
+                                                    {!! Form::open(['route'=>'role.edit','method'=>'get']) !!}
+                                                        {{ Form::hidden('id',$item->id) }}
+                                                        <button class="btn btn-xs btn-success">Edit</button>
 
-                                                {!! Form::close() !!}
+                                                    {!! Form::close() !!}
+                                                @endcan
+                                                @can('delete_role')
+                                                    {!! Form::open(['route'=>'role.delete','method'=>'delete']) !!}
+                                                        {{ Form::hidden('id',$item->id) }}
+                                                        <button class="btn btn-xs btn-danger">Delete</button>
 
-                                                {!! Form::open(['route'=>'role.delete','method'=>'delete']) !!}
-                                                    {{ Form::hidden('id',$item->id) }}
-                                                    <button class="btn btn-xs btn-danger">Delete</button>
-
-                                                {!! Form::close() !!}
+                                                    {!! Form::close() !!}
+                                                @endcan
                                             </td>
                                         </tr>
